fix: implement "loading state" for entity repository stores and only display components once loaded

+++ src/components/content/entity/recipe-file/list/RecipeFileList.svelte
@@ -1,6 +1,7 @@
 <script>
   import { recipeFileRepository } from "../../../../../services/repository/recipe-file-repository.ts";
   import { messages } from "../../../../../services/translation/en.ts";
+  import { isLoading } from "../../../../../services/util/is-loading.ts";
   import SvelteButton from "../../../../element/SvelteButton.svelte";
   import SvelteForm from "../../../../element/form/SvelteForm.svelte";
   import RecipeFileEdit from "../edit/RecipeFileEdit.svelte";
@@ -9,9 +10,10 @@
   export let recipeStepId;
 
   /**
-   * @type {Readable<number[]>}
+   * @type {Readable<Loadable<number[]>>}
    */
   let list;
+
   $: list = recipeFileRepository.createListFilteredStore({
     condition: { recipeStepId },
     orderBy: [
@@ -22,28 +24,30 @@
   });
 </script>
 
-<ol>
-  {#each $list as id}
-    <li>
-      <RecipeFileView id="{id}" /><SvelteButton
-        on:click="{() => recipeFileRepository.delete(id)}"
-        >{messages.labels.actions.delete.format()}</SvelteButton
-      >
-    </li>
-  {/each}
-</ol>
-<SvelteForm
-  on:submit="{async ({ detail: { values } }) => {
-    await recipeFileRepository.create({
-      name: values.name,
-      order: $list.length + 1,
-      path: values.path,
-      recipeStepId,
-    });
-  }}"
->
-  <RecipeFileEdit />
-  <SvelteButton type="submit"
-    >{messages.labels.actions.create.format()}</SvelteButton
+{#if !isLoading($list)}
+  <ol>
+    {#each $list as id}
+      <li>
+        <RecipeFileView id="{id}" /><SvelteButton
+          on:click="{() => recipeFileRepository.delete(id)}"
+          >{messages.labels.actions.delete.format()}</SvelteButton
+        >
+      </li>
+    {/each}
+  </ol>
+  <SvelteForm
+    on:submit="{async ({ detail: { values } }) => {
+      await recipeFileRepository.create({
+        name: values.name,
+        order: $list.length + 1,
+        path: values.path,
+        recipeStepId,
+      });
+    }}"
   >
-</SvelteForm>
+    <RecipeFileEdit />
+    <SvelteButton type="submit"
+      >{messages.labels.actions.create.format()}</SvelteButton
+    >
+  </SvelteForm>
+{/if}
